move @RumScreenName to otel package.
