+++ sample-app/src/main/java/com/splunk/android/sample/MainActivity.java
@@ -39,7 +39,7 @@ import androidx.navigation.Navigation;
 import androidx.navigation.ui.AppBarConfiguration;
 import androidx.navigation.ui.NavigationUI;
 import com.splunk.android.sample.databinding.ActivityMainBinding;
-import com.splunk.rum.RumScreenName;
+import io.opentelemetry.rum.internal.instrumentation.RumScreenName;
 import com.splunk.rum.SplunkRum;
 import io.opentelemetry.api.common.Attributes;
 import java.util.Arrays;
