+++ splunk-otel-android/src/main/java/io/opentelemetry/rum/internal/instrumentation/RumScreenName.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.splunk.rum;
+package io.opentelemetry.rum.internal.instrumentation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
