+++ splunk-otel-android/src/main/java/io/opentelemetry/rum/internal/instrumentation/fragment/FragmentTracer.java
@@ -19,10 +19,10 @@ package io.opentelemetry.rum.internal.instrumentation.fragment;
 import static io.opentelemetry.rum.internal.RumConstants.SCREEN_NAME_KEY;
 
 import androidx.fragment.app.Fragment;
-import com.splunk.rum.RumScreenName;
 import io.opentelemetry.api.common.AttributeKey;
 import io.opentelemetry.api.trace.Span;
 import io.opentelemetry.api.trace.Tracer;
+import io.opentelemetry.rum.internal.instrumentation.RumScreenName;
 import io.opentelemetry.rum.internal.util.ActiveSpan;
 
 class FragmentTracer {
